Conflicts:
	docs/pom.xml
	metrics-annotation/pom.xml
	metrics-benchmarks/pom.xml
	metrics-bom/pom.xml
	metrics-collectd/pom.xml
	metrics-collectd/src/test/java/io/dropwizard/metrics5/collectd/CollectdReporterTest.java
	metrics-core/pom.xml
	metrics-core/src/main/java/io/dropwizard/metrics5/MetricRegistry.java
	metrics-ehcache/pom.xml
	metrics-graphite/pom.xml
	metrics-healthchecks/pom.xml
	metrics-httpasyncclient/pom.xml
	metrics-httpclient/pom.xml
	metrics-httpclient5/pom.xml
	metrics-httpclient5/src/test/java/io/dropwizard/metrics5/httpclient5/InstrumentedHttpAsyncClientsTest.java
	metrics-httpclient5/src/test/java/io/dropwizard/metrics5/httpclient5/InstrumentedHttpClientsTest.java
	metrics-jcache/pom.xml
	metrics-jcstress/pom.xml
	metrics-jdbi/pom.xml
	metrics-jdbi3/pom.xml
	metrics-jersey2/pom.xml
	metrics-jetty9/pom.xml
	metrics-jetty9/src/main/java/io/dropwizard/metrics5/jetty9/InstrumentedHttpChannelListener.java
	metrics-jetty9/src/test/java/io/dropwizard/metrics5/jetty9/InstrumentedHttpChannelListenerTest.java
	metrics-jmx/pom.xml
	metrics-json/pom.xml
	metrics-jvm/pom.xml
	metrics-log4j2/pom.xml
	metrics-logback/pom.xml
	metrics-servlet/pom.xml
	metrics-servlets/pom.xml
	pom.xml

Refs https://github.com/dropwizard/metrics/issues/1735

+++ metrics-graphite/src/test/java/io/dropwizard/metrics5/graphite/PickledGraphiteTest.java
@@ -27,6 +27,7 @@ import javax.script.SimpleBindings;
 import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.math.BigInteger;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.Socket;
@@ -163,7 +164,7 @@ public class PickledGraphiteTest {
             bindings.put("payload", payload.substring(nextIndex));
             unpickleScript.eval(bindings);
             result.addAll(result.size(), (PyList) bindings.get("metrics"));
-            nextIndex += (Integer) bindings.get("batchLength");
+            nextIndex += ((BigInteger) bindings.get("batchLength")).intValue();
         }
 
         for (Object aResult : result) {
